+++ relay-server/src/actors/events.rs
@@ -42,9 +42,7 @@ use {
     crate::actors::store::{StoreEnvelope, StoreError, StoreForwarder},
     crate::service::ServerErrorKind,
     crate::utils::EnvelopeLimiter,
-    chrono::TimeZone,
     failure::ResultExt,
-    minidump::Minidump,
     relay_filter::FilterStatKey,
     relay_general::store::{GeoIpLookup, StoreConfig, StoreProcessor},
     relay_quotas::{DataCategory, RateLimitingError, RedisRateLimiter},
@@ -130,6 +128,9 @@ enum ProcessingError {
 
     #[fail(display = "envelope empty, transaction removed by sampling")]
     TransactionSampled,
+
+    #[fail(display = "envelope empty, event removed by sampling")]
+    EventSampled,
 }
 
 impl ProcessingError {
@@ -169,6 +170,7 @@ impl ProcessingError {
 
             // Dynamic sampling (not an error, just discarding messages that were removed by sampling)
             Self::TransactionSampled => Some(Outcome::Invalid(DiscardReason::TransactionSampled)),
+            Self::EventSampled => Some(Outcome::Invalid(DiscardReason::EventSampled)),
 
             // If we send to an upstream, we don't emit outcomes.
             Self::SendFailed(_) => None,
@@ -781,75 +783,6 @@ impl EventProcessor {
             .map_err(ProcessingError::InvalidUnrealReport)
     }
 
-    /// Writes a placeholder to indicate that this event has an associated minidump or an apple
-    /// crash report.
-    ///
-    /// This will indicate to the ingestion pipeline that this event will need to be processed. The
-    /// payload can be checked via `is_minidump_event`.
-    #[cfg(feature = "processing")]
-    fn write_native_placeholder(&self, event: &mut Event, is_minidump: bool) {
-        use relay_general::protocol::{Exception, JsonLenientString, Level, Mechanism};
-
-        // Events must be native platform.
-        let platform = event.platform.value_mut();
-        *platform = Some("native".to_string());
-
-        // Assume that this minidump is the result of a crash and assign the fatal
-        // level. Note that the use of `setdefault` here doesn't generally allow the
-        // user to override the minidump's level as processing will overwrite it
-        // later.
-        event.level.get_or_insert_with(|| Level::Fatal);
-
-        // Create a placeholder exception. This signals normalization that this is an
-        // error event and also serves as a placeholder if processing of the minidump
-        // fails.
-        let exceptions = event
-            .exceptions
-            .value_mut()
-            .get_or_insert_with(Values::default)
-            .values
-            .value_mut()
-            .get_or_insert_with(Vec::new);
-
-        exceptions.clear(); // clear previous errors if any
-
-        let (type_name, value, mechanism_type) = if is_minidump {
-            ("Minidump", "Invalid Minidump", "minidump")
-        } else {
-            (
-                "AppleCrashReport",
-                "Invalid Apple Crash Report",
-                "applecrashreport",
-            )
-        };
-
-        exceptions.push(Annotated::new(Exception {
-            ty: Annotated::new(type_name.to_string()),
-            value: Annotated::new(JsonLenientString(value.to_string())),
-            mechanism: Annotated::new(Mechanism {
-                ty: Annotated::from(mechanism_type.to_string()),
-                handled: Annotated::from(false),
-                synthetic: Annotated::from(true),
-                ..Mechanism::default()
-            }),
-            ..Exception::default()
-        }));
-    }
-
-    /// Extracts the timestamp from the minidump and uses it as the event timestamp.
-    #[cfg(feature = "processing")]
-    fn write_minidump_timestamp(&self, event: &mut Event, minidump_item: &Item) {
-        let minidump = match Minidump::read(minidump_item.payload()) {
-            Ok(minidump) => minidump,
-            Err(err) => {
-                relay_log::debug!("Failed to parse minidump: {:?}", err);
-                return;
-            }
-        };
-        let timestamp = Utc.timestamp(minidump.header.time_date_stamp.into(), 0);
-        event.timestamp.set_value(Some(timestamp.into()));
-    }
-
     /// Adds processing placeholders for special attachments.
     ///
     /// If special attachments are present in the envelope, this adds placeholder payloads to the
@@ -868,12 +801,11 @@ impl EventProcessor {
         if let Some(item) = minidump_attachment {
             let event = state.event.get_or_insert_with(Event::default);
             state.metrics.bytes_ingested_event_minidump = Annotated::new(item.len() as u64);
-            self.write_native_placeholder(event, true);
-            self.write_minidump_timestamp(event, item);
+            utils::process_minidump(event, &item.payload());
         } else if let Some(item) = apple_crash_report_attachment {
             let event = state.event.get_or_insert_with(Event::default);
             state.metrics.bytes_ingested_event_applecrashreport = Annotated::new(item.len() as u64);
-            self.write_native_placeholder(event, false);
+            utils::process_apple_crash_report(event, &item.payload());
         }
     }
 
@@ -1133,6 +1065,21 @@ impl EventProcessor {
         Ok(())
     }
 
+    /// Run dynamic sampling rules to see if we keep the event or remove it.
+    fn sample_event(&self, state: &mut ProcessEnvelopeState) -> Result<(), ProcessingError> {
+        let event = match &state.event.0 {
+            None => return Ok(()), // can't process without an event
+            Some(event) => event,
+        };
+
+        let project_id = state.project_id;
+        match utils::should_keep_event(event, &state.project_state, project_id) {
+            Some(false) => Err(ProcessingError::EventSampled),
+            Some(true) => Ok(()),
+            None => Ok(()), // Not enough info to make a definite evaluation, keep the event
+        }
+    }
+
     fn process_state(
         &self,
         mut state: ProcessEnvelopeState,
@@ -1164,6 +1111,8 @@ impl EventProcessor {
             });
 
             self.finalize_event(&mut state)?;
+
+            self.sample_event(&mut state)?;
 
             if_processing!({
                 self.store_process_event(&mut state)?;