+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 **Features**:
 
+- Support dynamic sampling for error events. ([#883](https://github.com/getsentry/relay/pull/883))
 - Remove `kafka-ssl` feature again because it breaks development workflow on OS X. ([#889](https://github.com/getsentry/relay/pull/889))
 
 **Bug Fixes**: